separate hmc utils
